[HAML-Lint] Improve issue ID and links for RuboCop
HAML-Lint uses internally RuboCop. But HAML-Lint issue's ID is always `RuboCop`.
This change converts it to `RuboCop:{cop_name}` such as `RuboCop:RuboCop:Lint/UselessAssignment`
and adds a link to the RuboCop cop documentation.

I think this change will allow users to handle HAML-Lint issues more easily.

Also, this change extracts common methods for RuboCop to a new module: `Runners::RuboCopUtils`.
