+++ lib/runners.rb
@@ -53,6 +53,7 @@ require "runners/ruby/gem_installer/source"
 require "runners/ruby/lockfile_loader"
 require "runners/ruby/lockfile_loader/lockfile"
 require "runners/ruby/lockfile_parser"
+require "runners/rubocop_utils"
 require "runners/nodejs"
 require "runners/nodejs/constraint"
 require "runners/nodejs/dependency"
