+++ test/smokes/haml_lint/expectations.rb
@@ -60,8 +60,11 @@ s.add_test(
   issues: [
     {
       message: "Lint/UselessAssignment: Useless assignment to variable - `unused_variable`.",
-      links: %W[https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop],
-      id: "RuboCop",
+      links: %W[
+        https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop
+        https://docs.rubocop.org/rubocop/cops_lint.html#lintuselessassignment
+      ],
+      id: "RuboCop:Lint/UselessAssignment",
       path: "test.haml",
       location: { start_line: 3 },
       object: { severity: "warning" },
@@ -78,8 +81,7 @@ s.add_test(
   type: "success",
   issues: [
     {
-      message:
-        "Lint/UselessAssignment: Useless assignment to variable - `unused_variable`. (https://rubystyle.guide#underscore-unused-vars)",
+      message: "Useless assignment to variable - `unused_variable`. (https://rubystyle.guide#underscore-unused-vars)",
       links: %w[https://github.com/sds/haml-lint/blob/v0.34.0/lib/haml_lint/linter#rubocop],
       id: "RuboCop",
       path: "test.haml",
@@ -111,9 +113,12 @@ s.add_test(
     {
       path: "test.haml",
       location: { start_line: 5 },
-      id: "RuboCop",
+      id: "RuboCop:Performance/FlatMap",
       message: "Performance/FlatMap: Use `flat_map` instead of `map...flatten`.",
-      links: %W[https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop],
+      links: %W[
+        https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop
+        https://docs.rubocop.org/rubocop-performance/cops_performance.html#performanceflatmap
+      ],
       object: { severity: "warning" },
       git_blame_info: {
         commit: :_, line_hash: "fbeff8480acd0bfaa22199218b67510fdfa9c389", original_line: 5, final_line: 5
