+++ test/smokes/haml_lint/with-inherit-gem/.rubocop.yml
@@ -1,3 +1,6 @@
 inherit_gem:
   meowcop:
     - "config/rubocop.yml"
+
+AllCops:
+  DisplayCopNames: false
